restructure middlewares and compose the middlewares framework
