+++ WeiboSpider/middlewares/ProxyMiddleware.py
@@ -1,0 +1,15 @@
+# -*- coding: utf-8 -*-
+# @Author  : CharesFuns
+# @Time    : 2021/7/18 19:13
+# @Function:
+
+class ProxyMiddleware(object):
+    def __init__(self):
+        pass
+
+    @classmethod
+    def from_crawler(cls, crawler):
+        pass
+
+    def process_request(self, request, crawler):
+        pass
