+++ WeiboSpider/middlewares/__init__.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+# @Author  : CharesFuns
+# @Time    : 2021/7/18 19:11
+# @Function:
+
+from WeiboSpider.middlewares.CheckMiddleware import CheckMiddleware
+from WeiboSpider.middlewares.ProxyMiddleware import ProxyMiddleware
+from WeiboSpider.middlewares.FakeUserAgentMiddleware import FakeUserAgentMiddleware
+
+__all__ = ['FakeUserAgentMiddleware', 'CheckMiddleware', 'ProxyMiddleware']
